backend/requirements.txt
altgraph==0.17.5
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Django==4.2.28
django-cors-headers==4.6.0
django-filter==24.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
idna==3.11
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
PyJWT==2.12.1
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
python-decouple==3.8
pywin32-ctypes==0.2.3
reportlab==4.4.9
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
